Split system test 2.2.delegate to improve stability & atomicity
